Tarefa6_2021li1gXXX

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2021/22.

Documentation

data RTree a

Constructors

R a [RTree a]

printCaminho :: IO ()

resolveJogo :: Int -> Jogo -> Maybe [Movimento]

A função resolveJogo iŕa receber um número de movimentos e um Jogo, e vai tentar resolver o jogo nesse número de movimentos

testPatchs :: [[Movimento]] -> Jogo -> Maybe [Movimento]

A função testPatchs vai pegar em todos os possíveis caminhos gerados pela função paths e vai aplicalos ao jogo, quando encontrar uma sequência de movimentos que levem o jogador até a porta, a função vai entregar essa lista de movimentos, caso não haja entrega Nothing

paths :: RTree a -> [[a]]

As funçoẽs paths e listRTree vão premetir ir a uma RoseTree (criada pela Função inicRtMov) e retirar desta todos os possíveis caminhos que o bot pode efetuar para um determinado número de movimentos

listRTree :: [a] -> RTree a -> [[a]]

daWin :: Jogo -> Bool

A Função daWin vai comparar as coordenadas do jogador com as coordenadas da porta,permite saber se o jogador já completou o jogo

inicRtMov :: Int -> Direcao -> RTree Movimento

A Função inicRtMov e a Função vão fucionar quase que como uma só, a inicRtMov vai dar inicio a uma RoseTree de Movimentos, com movimento para a esquerda , consequentemente o bot vai começar sempre a tentar andar para a esquerda

criaRtMov :: Direcao -> Int -> Movimento -> RTree Movimento

A Função criaRtMov vai dar continuidade a inicRtMov, vai construi uma RoseTree de Movimentos com uma profundidade de n (dado pela função resolveJogo)